ssss

+++ src/freequant/FreeQuant.QuantRouterLib/Data/CommandType.cs
@@ -1,12 +1,9 @@
-using System.Runtime.CompilerServices;
-
 namespace FreeQuant.QuantRouterLib.Data
 {
-  public class CommandType
-  {
-    public const byte Send = (byte) 0;
-    public const byte Cancel = (byte) 1;
-    public const byte Replace = (byte) 2;
-
-  }
+    public class CommandType
+    {
+        public const byte Send = 0;
+        public const byte Cancel = 1;
+        public const byte Replace = 2;
+    }
 }
